add "cover" (hero) image to single pages
I think it's a good idea to allow for images to precede posts to set the
tone and to capture the reader

cover is implemented in the same fashion as the figure shortcode, except
that cover uses front matter (via Params) instead of params to the
shortcode. a couple things are left out (like float) because we shouldn't
treat these as standalone figures.