spruce up item into photo "cards"
Like I want my posts to be memorable, the links to those posts should
also be memorable. Using the same "cover" photo for an item as with its
corresponding single should be an easy way to make my content more
recognizable. It's also a cool, grid-dy way to organize the page.

Implementation is very similar to the flexbox gallery items; the main
different here is that there are more overlays for the container:
0. background image
1. color overlay
2. text overlay
3. link overlay

This leaves us with a nice grid of clickable squares.

We're paginating by 6 because it looks good for both rows of 3 and 2.

`nonpost` is useful for albums and other things that don't quite act
like a "readable" blog.

`cover` frontmatter is used for both singles and their items.
`description` is a nice way of avoiding summaries.
Summaries are limited to 140 characters to prevent overflow.

If cover photo is not provided, a solid background is there in order to
maintain card-like appearance.

Cards are square so we're forcing height/width/padding-top 100%