Minor adjustment to padding around `.brand`
This gives even padding around a logo or text within `.brand`.  There's
a visual bug with text in which the right border isn't flush top to
bottom, but it looks fine with an image.

Displaying `a` as a block fills the `.brand` div, fixing a visual bug in
which spacing was inconsistent.