absURL for safer links; relative paths for figures
Instead of keeping images in the root `assets` directory, I would rather
keep them within leaf bundles instead. In order to make this easy, I'm
prepending the directory of the single onto any url referencing a figure
image (and its thumbnail). This can be turned by setting an `absolute`
parameter or key to `true`. This is the case for item covers, cover
header images, and figure shortcodes.

`absURL` is generally better than `.Site.BaseURL` because it's nice
about redundant forward slashes. In order to attempt to canonicalize all
links, we're going to use this liberally.

I've also ensured that taxonomies and their names are urlized in order
to deal with spaces (and special characters) nicely.