concat all javascript, remove CSP from meta
Our theme is not the place to set CSP headers; I personally will be
setting them with Lambda@Edge but I believe that this decision should be
left to the user. CSP was also breaking the `livereload.js` of hugo.

In order to compensate for this, I'm taking the simple way out and
hosting all the javascript dependencies myself. I'm also moving all
inline scripts into their own files and then concatenating them last.
That way, all of our scripts should be available under a single request.