Add fixes to make CSS work with tight CSP
Adding Content Security Policy headers broke inline styling for
galleries and image backgrounds. This isn't something we can stick in
our CSS files since the data varies based on the markdown. Our only
choice seems to be add "variable" styling via Javascript.

Since we're already using jQuery for `load-photoswipe.js`, we can use it
for our additional scripts. Eventually I would like to dump jQuery
because it's, by far, our biggest library and offers very little in
return.

We're wrapping `load-photoswipe.js` and `load-highlight.js` in IIFE's in
order to protect the global scope a little better.

We're also adding some protections against noscript with
`toggle-thumbnail-hidden.js` and `toggle-item-color-overlay.js`. The
adjustments in SCSS are to assist with that.